Add http_check e2e
