+++ datadog_checks_dev/datadog_checks/dev/_env.py
@@ -72,27 +72,29 @@ def format_config(config):
 
 
 def replay_check_run(agent_collector, stub_aggregator):
-    # We assume a test only uses one instance
-    agent_collector = agent_collector[0]
-    aggregator = agent_collector['aggregator']
-    runner = agent_collector['runner']
-    check_id = runner['CheckID']
-    check_name = runner['CheckName']
+    errors = []
+    for collector in agent_collector:
+        aggregator = collector['aggregator']
+        runner = collector['runner']
+        check_id = runner['CheckID']
+        check_name = runner['CheckName']
 
-    for data in aggregator.get('metrics', []):
-        for _, value in data['points']:
-            metric_type = stub_aggregator.METRIC_ENUM_MAP[data['type']]
-            stub_aggregator.submit_metric(
-                check_name, check_id, metric_type, data['metric'], value, data['tags'], data['host']
+        for data in aggregator.get('metrics', []):
+            for _, value in data['points']:
+                metric_type = stub_aggregator.METRIC_ENUM_MAP[data['type']]
+                stub_aggregator.submit_metric(
+                    check_name, check_id, metric_type, data['metric'], value, data['tags'], data['host']
+                )
+
+        for data in aggregator.get('service_checks', []):
+            stub_aggregator.submit_service_check(
+                check_name, check_id, data['check'], data['status'], data['tags'], data['host_name'], data['message']
             )
 
-    for data in aggregator.get('service_checks', []):
-        stub_aggregator.submit_service_check(
-            check_name, check_id, data['check'], data['status'], data['tags'], data['host_name'], data['message']
-        )
-
-    if runner['LastError']:
-        raise Exception(runner['LastError'])
+        if runner['LastError']:
+            errors.append(runner['LastError'])
+    if errors:
+        raise Exception(errors)
 
 
 def serialize_data(data):
