+++ http_check/tests/test_unit.py
@@ -1,0 +1,24 @@
+# -*- coding: utf-8 -*-
+
+# (C) Datadog, Inc. 2018
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+
+import mock
+
+from datadog_checks.http_check import HTTPCheck
+
+
+def test__init__():
+    # empty values should be ignored
+    init_config = {'ca_certs': ''}
+    # `get_ca_certs_path` needs to be mocked because it's used as fallback when
+    # init_config doesn't contain `ca_certs`
+    with mock.patch('datadog_checks.http_check.http_check.get_ca_certs_path', return_value='bar'):
+        http_check = HTTPCheck('http_check', init_config, {})
+        assert http_check.ca_certs == 'bar'
+
+    # normal case
+    init_config = {'ca_certs': 'foo'}
+    http_check = HTTPCheck('http_check', init_config, {})
+    assert http_check.ca_certs == 'foo'
